rename re_index to reindex
