+++ lib/collection.js
@@ -57,7 +57,7 @@ const ListIndexesOperation = require('./operations/list_indexes');
 const MapReduceOperation = require('./operations/map_reduce');
 const OptionsOperation = require('./operations/options_operation');
 const RenameOperation = require('./operations/rename');
-const ReIndexOperation = require('./operations/re_index');
+const ReIndexOperation = require('./operations/reindex');
 const ReplaceOneOperation = require('./operations/replace_one');
 const StatsOperation = require('./operations/stats');
 const UpdateManyOperation = require('./operations/update_many');
