+++ lib/operations/reindex.js
